% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp.R
\name{revcomp}
\alias{revcomp}
\title{Reverse compliment}
\usage{
revcomp(seq)
}
\arguments{
\item{seq}{A nucleotide sequence.}
}
\value{
A character vector
}
\description{
Reverse compliment
}
\examples{
revcomp("AGTCG")
}
